Moved all server models to a central models folder. Started to add recovery to new user payload for #98.
