+++ src/client/app/user/user-register/user-register.component.html
@@ -12,10 +12,9 @@
     <div class="custTable">
       <div class="row">
         <div class="cell input">
-          <mat-form-field>
+          <mat-form-field class="user">
             <input
               matInput
-              class="user"
               [(ngModel)]="username"
               type="email"
               placeholder="Email"
@@ -25,10 +24,9 @@
       </div>
       <div class="row">
         <div class="cell input">
-          <mat-form-field>
+          <mat-form-field class="user">
             <input
               matInput
-              class="user"
               [(ngModel)]="password"
               type="password"
               placeholder="Password"
@@ -38,10 +36,9 @@
       </div>
       <div class="row">
         <div class="cell input">
-          <mat-form-field>
+          <mat-form-field class="user">
             <input
               matInput
-              class="user"
               [(ngModel)]="passwordConfirmation"
               type="password"
               placeholder="Confirm Password"
