+++ src/server/app/auth/auth.pipe.ts
@@ -6,7 +6,7 @@ import {
   ValidationPipe
 } from '@nestjs/common';
 
-import { NewUserDTO } from '@Auth/interfaces/new_user.dto';
+import { NewUserDTO } from '@Models/auth/new_user.dto';
 
 @Injectable()
 export class AuthPipe extends ValidationPipe
