+++ src/server/app/models/auth/new_user.dto.ts
@@ -1,5 +1,7 @@
 import { ApiModelProperty } from '@nestjs/swagger';
 
+import { Recovery } from './recovery.dto';
+
 export class NewUserDTO {
   @ApiModelProperty()
   readonly email: string;
@@ -7,4 +9,6 @@ export class NewUserDTO {
   readonly password: string;
   @ApiModelProperty()
   readonly confirmationPassword: string;
+  @ApiModelProperty({ type: [Recovery] })
+  readonly recovery: Recovery[];
 }
