+++ src/server/app/models/auth/recovery.dto.ts
@@ -1,0 +1,8 @@
+import { ApiModelProperty } from '@nestjs/swagger';
+
+export class Recovery {
+  @ApiModelProperty()
+  readonly question: string;
+  @ApiModelProperty()
+  readonly answer: string;
+}
