+++ src/server/app/models/character/attribute.dto.ts
