+++ src/server/app/models/character/save.dto.ts
