+++ src/server/app/models/note/note.dto.ts
