+++ src/server/app/models/skill/skill.dto.ts
