+++ src/server/app/models/spell/spell.dto.ts
