+++ src/server/app/models/weapon/weapon.dto.ts
