+++ src/server/app/zeldaplay/character/character.pipe.ts
@@ -1,11 +1,11 @@
 import { ArgumentMetadata, Injectable, PipeTransform } from '@nestjs/common';
 
-import { CharacterDTO } from '@Character/interfaces/character.dto';
 import { DbCharacter } from '@DbModel/db_character.model';
 import { DbSave } from '@DbModel/db_save.model';
 import { DbSkill } from '@DbModel/db_skill.model';
 import { Modifier } from '@DbModel/enums/modifiers.enum';
 import { Race } from '@DbModel/enums/races.enum';
+import { CharacterDTO } from '@Models/character/character.dto';
 
 @Injectable()
 export class CharacterPipe implements PipeTransform<CharacterDTO, DbCharacter> {
