+++ src/server/app/zeldaplay/spell/spell.pipe.spec.ts
@@ -1,6 +1,6 @@
 import { DbSpell } from '@DbModel/db_spell.table';
 import { SpellPipe } from './spell.pipe';
-import { SpellDTO } from './interfaces/spell.dto';
+import { SpellDTO } from '../../models/spell/spell.dto';
 import { Modifier } from '@Db/models/enums/modifiers.enum';
 
 describe('#SpellPipe', () => {
